requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
SpeechRecognition>=3.10.0
requests>=2.31.0
numpy>=1.24.0
PyAudio>=0.2.13
python-magic>=0.4.27
google-cloud-core
google-auth
google-api-python-client
google-cloud-aiplatform
python-docx==1.1.0
PyPDF2==3.0.1
python-multipart==0.0.9
Pillow==10.0.0
